+++ pyramid/__init__.py
@@ -4,7 +4,9 @@
 #
 # The pyramid module
 
-__version__ = "0.8.1"
+import os as _os
+
+__version__ = "0.9.0"
 
 try:
     # this var is injected in the setup build to enable
@@ -16,9 +18,8 @@ except NameError:
 
 if __PYRAMID_SETUP__:
     import sys
-    import os
     sys.stderr.write('Partial import of pyramid during the build process.' +
-                     os.linesep)
+                     _os.linesep)
 else:
     # check that the build completed properly. This prints an informative
     # message in the case that any of the C code was not properly compiled.
@@ -52,7 +53,29 @@ else:
         'plot_pacf'
     ]
 
+    # Warn for Issue #34 (https://github.com/tgsmith61591/pyramid/issues/34)
+    import warnings
+    warnings.warn("""
+    The 'pyramid' package will be migrating to a new namespace beginning in 
+    version 1.0.0: 'pmdarima'. This is due to a package name collision with the
+    Pyramid web framework. For more information, see Issue #34:
+    
+        https://github.com/tgsmith61591/pyramid/issues/34
+        
+    The package will subsequently be installable via the name 'pmdarima'; the
+    only functional change to the user will be the import name. All imports
+    from 'pyramid' will change to 'pmdarima'.
+    """, UserWarning)
+
+    # On first import, check the cache, warn if needed
+    from ._config import _warn_for_cache_size
+    _warn_for_cache_size()
+
     # Delete unwanted variables from global
+    del _os
+    del warnings
+    # del _config  # don't delete in case user wants to amend it at top level
+    del _warn_for_cache_size
     del __check_build
     del __PYRAMID_SETUP__