+++ pyramid/arima/tests/test_arima.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 
 from pyramid.arima import ARIMA, auto_arima
 from pyramid.arima.arima import VALID_SCORING
-from pyramid.arima.auto import _fmt_warning_str
+from pyramid.arima.auto import _fmt_warning_str, _post_ppc_arima
 from pyramid.arima.utils import nsdiffs
 from pyramid.datasets import load_lynx, load_wineind, load_heartrate
 from pyramid.utils import get_callable, assert_raises
@@ -13,11 +13,15 @@ import numpy as np
 from numpy.testing import assert_array_almost_equal, assert_almost_equal
 from numpy.random import RandomState
 
+from sklearn.externals import joblib
+
 from statsmodels import api as sm
 import pandas as pd
 
 import warnings
 import pickle
+import pytest
+import time
 import os
 
 # initialize the random state
@@ -26,11 +30,6 @@ y = rs.rand(25)
 
 # more interesting heart rate data (asserts we can use a series)
 hr = load_heartrate(as_series=True)
-
-# Yes, m is ACTUALLY 12... but that takes a LONG time. If we set it to
-# 1, we actually get a much, much faster model fit. We can only use this
-# if we're NOT testing the output of the model, but just the functionality!
-wineind_m = 1
 
 # > set.seed(123)
 # > abc <- rnorm(50, 5, 1)
@@ -50,6 +49,14 @@ abc = np.array([4.439524, 4.769823, 6.558708, 5.070508,
 
 wineind = load_wineind()
 lynx = load_lynx()
+
+# Yes, m is ACTUALLY 12... but that takes a LONG time. If we set it to
+# 1, we actually get a much, much faster model fit. We can only use this
+# if we're NOT testing the output of the model, but just the functionality!
+wineind_m = 1
+
+# A random xreg for the wineind array
+wineind_xreg = rs.rand(wineind.shape[0], 2)
 
 
 def test_basic_arima():
@@ -169,7 +176,7 @@ def test_oob_for_issue_28():
 
     # Show we fail if we try to add observations with a different dim exog
     assert_raises(ValueError, arima_no_oob.add_new_observations,
-                  hr[-10:], xreg_test[:, 2])
+                  hr[-10:], xreg_test[:, :2])
 
     # Actually add them now, and compare the forecasts (should be the same)
     arima_no_oob.add_new_observations(hr[-10:], xreg[-10:, :])
@@ -391,7 +398,8 @@ def test_errors():
     # show errors for d
     _assert_val_error(auto_arima, abc, max_d=-1)
     _assert_val_error(auto_arima, abc, d=-1)
-    _assert_val_error(auto_arima, abc, d=5, max_d=4)
+    _assert_val_error(auto_arima, abc, max_D=-1)
+    _assert_val_error(auto_arima, abc, D=-1)
 
     # show error for bad IC
     _assert_val_error(auto_arima, abc, information_criterion='bad-value')
@@ -456,7 +464,7 @@ def test_with_seasonality2():
     def do_fit():
         return auto_arima(wineind, start_p=1, start_q=1, max_p=2,
                           max_q=2, m=2, start_P=0,
-                          seasonal=True, n_jobs=-1,
+                          seasonal=True, n_jobs=2,
                           d=1, D=1, stepwise=False,
                           suppress_warnings=True,
                           error_action='ignore',
@@ -469,6 +477,9 @@ def test_with_seasonality2():
 
     # ensure summary still works
     seasonal_fit.summary()
+
+    # Show we can predict on seasonal where conf_int is true
+    seasonal_fit.predict(n_periods=10, return_conf_int=True)
 
 
 def test_with_seasonality3():
@@ -485,7 +496,7 @@ def test_with_seasonality4():
     # make the function return all the values. Also, use small M to make our
     # lives easier.
     auto_arima(wineind, start_p=1, start_q=1, max_p=2, max_q=2, m=12,
-               start_P=0, seasonal=True, n_jobs=2, d=1, D=None, stepwise=False,
+               start_P=0, seasonal=True, n_jobs=1, d=1, D=None, stepwise=False,
                error_action='ignore', suppress_warnings=True,
                random=True, random_state=42, return_valid_fits=True,
                n_fits=3)  # only a few
@@ -494,12 +505,13 @@ def test_with_seasonality4():
 def test_with_seasonality5():
     # can we fit the same thing with an exogenous array of predictors?
     # also make it stationary and make sure that works...
+    # 9/22/18 - make not parallel to reduce mem overhead on pytest
     all_res = auto_arima(wineind, start_p=1, start_q=1, max_p=2,
-                         max_q=2, m=12, start_P=0, seasonal=True, n_jobs=2,
+                         max_q=2, m=12, start_P=0, seasonal=True,
                          d=1, D=None, error_action='ignore',
                          suppress_warnings=True, stationary=True,
-                         random=True, random_state=42, return_valid_fits=True,
-                         stepwise=False, n_fits=3,  # only a few
+                         random_state=42, return_valid_fits=True,
+                         stepwise=True,
                          exogenous=rs.rand(wineind.shape[0], 4))  # only fit 2
 
     # show it is a list
@@ -522,7 +534,7 @@ def test_with_seasonality7():
                out_of_sample_size=10, information_criterion='oob',
                suppress_warnings=True,
                error_action='raise',  # do raise so it fails fast
-               random=True, random_state=42, n_fits=3,
+               random=True, random_state=42, n_fits=2,
                stepwise=False)
 
 
@@ -558,3 +570,187 @@ def test_warning_str_fmt():
 
 def test_nsdiffs_on_wine():
     assert nsdiffs(wineind, m=52) == 2
+
+
+# Asserting where D grows too large as a product of an M that's too big.
+def test_m_too_large():
+    train = lynx[:90]
+
+    with pytest.raises(ValueError) as v:
+        auto_arima(train, start_p=1, start_q=1, start_P=1, start_Q=1,
+                   max_p=5, max_q=5, max_P=5, max_Q=5, seasonal=True,
+                   stepwise=True, suppress_warnings=True, D=10, max_D=10,
+                   error_action='ignore', m=20)
+
+        msg = str(v)
+        assert 'The seasonal differencing order' in msg
+
+
+# Test that (as of v0.9.1) we can pickle a model, pickle it again, load both
+# and create predictions.
+def test_double_pickle():
+    arima = ARIMA(order=(0, 0, 0), trend='c', suppress_warnings=True)
+    arima.fit(y)
+
+    # Now save it twice
+    file_a = 'first.pkl'
+    file_b = 'second.pkl'
+
+    try:
+        # No compression
+        joblib.dump(arima, file_a)
+
+        # Sleep between pickling so that the "pickle hash" for the ARIMA is
+        # different by enough. We could theoretically also just use a UUID
+        # for part of the hash to make sure it's unique?
+        time.sleep(0.5)
+
+        # Some compression
+        joblib.dump(arima, file_b, compress=2)
+
+        # Load both and prove they can both predict
+        loaded_a = joblib.load(file_a)  # type: ARIMA
+        loaded_b = joblib.load(file_b)  # type: ARIMA
+        pred_a = loaded_a.predict(n_periods=5)
+        pred_b = loaded_b.predict(n_periods=5)
+        assert np.allclose(pred_a, pred_b)
+
+        # Remove the caches from each
+        loaded_a._clear_cached_state()
+        loaded_b._clear_cached_state()
+
+        # Test the previous condition where we removed the saved state of an
+        # ARIMA from statsmodels and caused an OSError and a corrupted pickle
+        with pytest.raises(OSError) as o:
+            joblib.load(file_a)  # fails since no cached state there!
+            msg = str(o)
+            assert 'Could not read saved model state' in msg, msg
+
+    # Always remove in case we fail in try, leaving residual files
+    finally:
+        os.unlink(file_a)
+        os.unlink(file_b)
+
+
+# We fail if we don't end up fitting any models in the auto_arima func
+def test_value_error_on_failed_model_fits():
+    with pytest.raises(ValueError):
+        _post_ppc_arima(None)
+
+
+# We fail when error action is raise and the model can't be fit
+def test_failing_model_fit():
+    with pytest.raises(ValueError):
+        # raise ValueError('non-invertible starting MA parameters found'
+        auto_arima(wineind, seasonal=True, suppress_warnings=True,
+                   error_action='raise', m=2, random=True, random_state=1,
+                   n_fits=2)
+
+
+def test_warn_for_large_differences():
+    # First: d is too large
+    with warnings.catch_warnings(record=True) as w:
+        _ = auto_arima(wineind, seasonal=True, m=1, suppress_warnings=False,
+                       d=3, error_action='warn')
+
+        assert len(w) > 0
+
+    # Second: D is too large. M needs to be > 1 or D will be set to 0...
+    # unfortunately, this takes a long time.
+    with warnings.catch_warnings(record=True) as w:
+        _ = auto_arima(wineind, seasonal=True, m=2, suppress_warnings=False,
+                       D=3, error_action='warn')
+
+        assert len(w) > 0
+
+
+def test_warn_for_stepwise_and_parallel():
+    with warnings.catch_warnings(record=True) as w:
+        _ = auto_arima(lynx, suppress_warnings=False, d=1,
+                       error_action='ignore', stepwise=True, n_jobs=2)
+
+        assert len(w) > 0
+
+
+# Force case where data is simple polynomial after differencing
+def test_force_polynomial_error():
+    x = np.array([1, 2, 3, 4, 5, 6])
+    d = 2
+    xreg = None
+
+    with pytest.raises(ValueError) as ve:
+        auto_arima(x, d=d, D=0, seasonal=False, exogenous=xreg)
+        assert 'simple polynomial' in str(ve), str(ve)
+
+    # but it should pass when xreg is not none
+    xreg = rs.rand(x.shape[0], 2)
+    _ = auto_arima(x, d=d, D=0, seasonal=False, exogenous=xreg,
+                   error_action='ignore', suppress_warnings=True)
+
+
+# Test if exogenous is not None and D > 0
+def test_seasonal_xreg_differencing():
+    # Test both a small M and a large M since M is used as the lag parameter
+    # in the xreg array differencing. If M is 1, D is set to 0
+    for m in (2,):  # 12): takes FOREVER
+        _ = auto_arima(wineind, d=1, D=1, seasonal=True,
+                       exogenous=wineind_xreg, error_action='ignore',
+                       suppress_warnings=True, m=m)
+
+
+# Show that we can complete when max order is None
+def test_inf_max_order():
+    _ = auto_arima(lynx, max_order=None, suppress_warnings=True,
+                   error_action='ignore')
+
+
+# Regression testing for unpickling an ARIMA from an older version
+def test_for_older_version():
+    # Fit an ARIMA
+    arima = ARIMA(order=(0, 0, 0), trend='c', suppress_warnings=True)
+
+    # There are three possibilities here:
+    # 1. The model is serialized/deserialized BEFORE it has been fit.
+    #    This means we should not get a warning.
+    #
+    # 2. The model is saved after being fit, but it does not have a
+    #    pkg_version_ attribute due to it being an old (very old) version.
+    #    We still warn for this
+    #
+    # 3. The model is saved after the fit, and it's version does not match.
+    #    We warn for this.
+    for case, do_fit, expect_warning in [(1, False, False),
+                                         (2, True, True),
+                                         (3, True, True)]:
+
+        # Only fit it if we should
+        if do_fit:
+            arima.fit(y)
+
+        # If it's case 2, we remove the pkg_version_. If 3, we set it low
+        if case == 2:
+            delattr(arima, 'pkg_version_')
+        elif case == 3:
+            arima.pkg_version_ = '0.0.1'  # will always be < than current
+
+        # Pickle it
+        pickle_file = 'model.pkl'
+        try:
+            joblib.dump(arima, pickle_file)
+
+            # Now unpickle it and show that we get a warning (if expected)
+            with warnings.catch_warnings(record=True) as w:
+                arm = joblib.load(pickle_file)  # type: ARIMA
+
+                if expect_warning:
+                    assert len(w) > 0
+                else:
+                    assert not len(w)
+
+                # we can still produce predictions (only if we fit)
+                if do_fit:
+                    arm.predict(n_periods=4)
+
+        finally:
+            arima._clear_cached_state()
+            os.unlink(pickle_file)