+++ pyramid/compat/__init__.py
@@ -9,5 +9,6 @@ from .matplotlib import *
 from .pandas import *
 from .python import *
 from .numpy import *
+from .statsmodels import *
 
 __all__ = [s for s in dir() if not s.startswith('_')]